# admin/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AbitaDash API",
        "version": "1.0.0",
        "description": "Public API for AbitaDash - Categories and Products"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "summary": "Get all categories",
                "description": "Retrieve all product categories for the home page",
                "tags": ["Categories"],
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": { "type": "integer" },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "slug": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "imageUrl": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "summary": "Get all products",
                "description": "Retrieve all products",
                "tags": ["Products"],
                "responses": {
                    "200": {
                        "description": "List of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": { "type": "integer" },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "slug": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "imageUrl": {
                                                        "type": "string"
                                                    },
                                                    "categoryId": {
                                                        "type": "integer"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categories/{categorySlug}/products": {
            "get": {
                "summary": "Get products by category",
                "description": "Retrieve all products for a specific category by its slug",
                "tags": ["Products"],
                "parameters": [
                    {
                        "name": "categorySlug",
                        "in": "path",
                        "description": "The category slug",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products in category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": { "type": "integer" },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "slug": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "imageUrl": {
                                                        "type": "string"
                                                    },
                                                    "categoryId": {
                                                        "type": "integer"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
        "/products/{productSlug}": {
            "get": {
                "summary": "Get product with related products",
                "description": "Retrieve a single product by slug with 3 random related products from the same category",
                "tags": ["Products"],
                "parameters": [
                    {
                        "name": "productSlug",
                        "in": "path",
                        "description": "The product slug",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details with related products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": { "type": "integer" },
                                                "name": { "type": "string" },
                                                "slug": { "type": "string" },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "imageUrl": {
                                                    "type": "string"
                                                },
                                                "categoryId": {
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        "relatedProducts": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": { "type": "integer" },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "slug": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "imageUrl": {
                                                        "type": "string"
                                                    },
                                                    "categoryId": {
                                                        "type": "integer"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/quotes": {
            "post": {
                "summary": "Create a quote",
                "description": "Create a new quote request",
                "tags": ["Quotes"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "companyName",
                                    "contactName",
                                    "email",
                                    "phone",
                                    "description"
                                ],
                                "properties": {
                                    "companyName": { "type": "string" },
                                    "contactName": { "type": "string" },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": { "type": "string" },
                                    "description": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Quote created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": { "type": "string" },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": { "type": "integer" },
                                                "companyName": {
                                                    "type": "string"
                                                },
                                                "contactName": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email"
                                                },
                                                "phone": { "type": "string" },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "createdAt": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/contact-messages": {
            "post": {
                "summary": "Create a contact message",
                "description": "Create a new contact message from client-side contact form",
                "tags": ["Messages"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "fullName",
                                    "companyName",
                                    "phone",
                                    "email",
                                    "message"
                                ],
                                "properties": {
                                    "fullName": { "type": "string" },
                                    "companyName": { "type": "string" },
                                    "phone": { "type": "string" },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "message": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact message created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": { "type": "string" },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": { "type": "integer" },
                                                "fullName": {
                                                    "type": "string"
                                                },
                                                "companyName": {
                                                    "type": "string"
                                                },
                                                "phone": { "type": "string" },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email"
                                                },
                                                "message": { "type": "string" },
                                                "createdAt": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        }
    }
}